requirements.txt
streamlit==1.51.0
PyPDF2==3.0.1
python-dotenv==1.2.1
langchain==1.1.0
langchain-openai==1.1.0
langchain-community==0.4.1
langchain-text-splitters==1.0.0
langchain-core==1.1.0
faiss-cpu==1.13.0
sentence-transformers==5.1.2
InstructorEmbedding==1.0.1
torch==2.9.1
